# perf: opossum