--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -5,18 +5,21 @@ const Status = require('./status');
 const Semaphore = require('./semaphore');
 const MemoryCache = require('./cache');
 
+const kStateOpen = 1 << 1;
+const kStateClosed = 1 << 2;
+const kStateHalfOpen = 1 << 3;
+const kStatePendingClose = 1 << 4;
+const kStateShutdown = 1 << 5;
+const kStateWarmingUp = 1 << 6;
+const kStateEnabled = 1 << 7;
+
+const STATE_BITSET = Symbol('stateBitset');
 const STATE = Symbol('state');
-const OPEN = Symbol('open');
-const CLOSED = Symbol('closed');
-const HALF_OPEN = Symbol('half-open');
-const PENDING_CLOSE = Symbol('pending-close');
-const SHUTDOWN = Symbol('shutdown');
+
 const FALLBACK_FUNCTION = Symbol('fallback');
 const STATUS = Symbol('status');
 const NAME = Symbol('name');
 const GROUP = Symbol('group');
-const ENABLED = Symbol('Enabled');
-const WARMING_UP = Symbol('warming-up');
 const VOLUME_THRESHOLD = Symbol('volume-threshold');
 const OUR_ERROR = Symbol('our-error');
 const RESET_TIMEOUT = Symbol('reset-timeout');
@@ -24,6 +27,10 @@ const WARMUP_TIMEOUT = Symbol('warmup-timeout');
 const LAST_TIMER_AT = Symbol('last-timer-at');
 const deprecation = `options.maxFailures is deprecated. \
 Please use options.errorThresholdPercentage`;
+
+function hasFlag (object, flag) {
+  return (object[STATE_BITSET] & flag) > 0;
+}
 
 /**
  * Constructs a {@link CircuitBreaker}.
@@ -202,7 +209,9 @@ class CircuitBreaker extends EventEmitter {
     this[VOLUME_THRESHOLD] = Number.isInteger(options.volumeThreshold)
       ? options.volumeThreshold
       : 0;
-    this[WARMING_UP] = options.allowWarmUp === true;
+    this[STATE_BITSET] = 0;
+
+    if (options.allowWarmUp === true) { this[STATE_BITSET] |= kStateWarmingUp; }
 
     // The user can pass in a Status object to initialize the Status/stats
     if (this.options.status) {
@@ -216,31 +225,49 @@ class CircuitBreaker extends EventEmitter {
       this[STATUS] = new Status(this.options);
     }
 
-    this[STATE] = CLOSED;
+    this[STATE] = kStateClosed;
 
     if (options.state) {
-      this[ENABLED] = options.state.enabled !== false;
-      this[WARMING_UP] = options.state.warmUp || this[WARMING_UP];
+      if (options.state.enabled !== false) {
+        this[STATE_BITSET] |= kStateEnabled;
+      }
+
+      if (options.state.warmUp || hasFlag(this, kStateWarmingUp)) {
+        this[STATE_BITSET] |= kStateWarmingUp;
+      }
+
       // Closed if nothing is passed in
-      this[CLOSED] = options.state.closed !== false;
+      if (options.state.closed !== false) {
+        this[STATE_BITSET] |= kStateClosed;
+      }
+
       // These should be in sync
-      this[HALF_OPEN] = this[PENDING_CLOSE] = options.state.halfOpen || false;
+      if (options.state.halfOpen) {
+        this[STATE_BITSET] |= kStateHalfOpen | kStatePendingClose;
+      }
+
       // Open should be the opposite of closed,
       // but also the opposite of half_open
-      this[OPEN] = !this[CLOSED] && !this[HALF_OPEN];
-      this[SHUTDOWN] = options.state.shutdown || false;
+      if (!hasFlag(this, kStateClosed) && !hasFlag(this, kStateHalfOpen)) {
+        this[STATE_BITSET] |= kStateOpen;
+      }
+
+      if (options.state.shutdown) {
+        this[STATE_BITSET] |= kStateShutdown;
+      }
     } else {
-      this[PENDING_CLOSE] = false;
-      this[ENABLED] = options.enabled !== false;
+      if (options.enabled !== false) {
+        this[STATE_BITSET] |= kStateEnabled;
+      }
     }
 
     this[FALLBACK_FUNCTION] = null;
     this[NAME] = options.name || action.name || nextName();
     this[GROUP] = options.group || this[NAME];
 
-    if (this[WARMING_UP]) {
+    if (hasFlag(this, kStateWarmingUp)) {
       const timer = this[WARMUP_TIMEOUT] = setTimeout(
-        _ => (this[WARMING_UP] = false),
+        _ => (this[STATE_BITSET] &= ~kStateWarmingUp),
         this.options.rollingCountTimeout
       );
       if (typeof timer.unref === 'function') {
@@ -293,8 +320,8 @@ class CircuitBreaker extends EventEmitter {
      * @returns {void}
      */
     function _halfOpen (circuit) {
-      circuit[STATE] = HALF_OPEN;
-      circuit[PENDING_CLOSE] = true;
+      circuit[STATE] = kStateHalfOpen;
+      circuit[STATE_BITSET] |= kStatePendingClose;
       /**
        * Emitted after `options.resetTimeout` has elapsed, allowing for
        * a single attempt to call the service again. If that attempt is
@@ -314,12 +341,12 @@ class CircuitBreaker extends EventEmitter {
     });
 
     // Prepopulate the State of the Breaker
-    if (this[SHUTDOWN]) {
-      this[STATE] = SHUTDOWN;
+    if (hasFlag(this, kStateShutdown)) {
+      this[STATE] = kStateShutdown;
       this.shutdown();
-    } else if (this[CLOSED]) {
+    } else if (hasFlag(this, kStateClosed)) {
       this.close();
-    } else if (this[OPEN]) {
+    } else if (hasFlag(this, kStateOpen)) {
       // If the state being passed in is OPEN but more time has elapsed
       // than the resetTimeout, then we should be in halfOpen state
       if (this.options.state.lastTimerAt !== undefined &&
@@ -329,9 +356,9 @@ class CircuitBreaker extends EventEmitter {
       } else {
         this.open();
       }
-    } else if (this[HALF_OPEN]) {
+    } else if (hasFlag(this, kStateHalfOpen)) {
       // Not sure if anything needs to be done here
-      this[STATE] = HALF_OPEN;
+      this[STATE] = kStateHalfOpen;
     }
   }
 
@@ -341,12 +368,12 @@ class CircuitBreaker extends EventEmitter {
    * @returns {void}
    */
   close () {
-    if (this[STATE] !== CLOSED) {
+    if (this[STATE] !== kStateClosed) {
       if (this[RESET_TIMEOUT]) {
         clearTimeout(this[RESET_TIMEOUT]);
       }
-      this[STATE] = CLOSED;
-      this[PENDING_CLOSE] = false;
+      this[STATE] = kStateClosed;
+      this[STATE_BITSET] &= ~kStatePendingClose;
       /**
        * Emitted when the breaker is reset allowing the action to execute again
        * @event CircuitBreaker#close
@@ -365,9 +392,9 @@ class CircuitBreaker extends EventEmitter {
    * @returns {void}
    */
   open () {
-    if (this[STATE] !== OPEN) {
-      this[STATE] = OPEN;
-      this[PENDING_CLOSE] = false;
+    if (this[STATE] !== kStateOpen) {
+      this[STATE] = kStateOpen;
+      this[STATE_BITSET] &= ~kStatePendingClose;
       /**
        * Emitted when the breaker opens because the action has
        * failure percentage greater than `options.errorThresholdPercentage`.
@@ -398,7 +425,7 @@ class CircuitBreaker extends EventEmitter {
       clearTimeout(this[WARMUP_TIMEOUT]);
     }
     this.status.shutdown();
-    this[STATE] = SHUTDOWN;
+    this[STATE] = kStateShutdown;
 
     // clear cache on shutdown
     this.clearCache();
@@ -409,7 +436,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get isShutdown () {
-    return this[STATE] === SHUTDOWN;
+    return this[STATE] === kStateShutdown;
   }
 
   /**
@@ -433,7 +460,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get pendingClose () {
-    return this[PENDING_CLOSE];
+    return hasFlag(this, kStatePendingClose);
   }
 
   /**
@@ -441,7 +468,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get closed () {
-    return this[STATE] === CLOSED;
+    return this[STATE] === kStateClosed;
   }
 
   /**
@@ -449,7 +476,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get opened () {
-    return this[STATE] === OPEN;
+    return this[STATE] === kStateOpen;
   }
 
   /**
@@ -457,7 +484,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get halfOpen () {
-    return this[STATE] === HALF_OPEN;
+    return this[STATE] === kStateHalfOpen;
   }
 
   /**
@@ -498,7 +525,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get enabled () {
-    return this[ENABLED];
+    return hasFlag(this, kStateEnabled);
   }
 
   /**
@@ -506,7 +533,7 @@ class CircuitBreaker extends EventEmitter {
    * @type {Boolean}
    */
   get warmUp () {
-    return this[WARMING_UP];
+    return hasFlag(this, kStateWarmingUp);
   }
 
   /**
@@ -624,7 +651,7 @@ class CircuitBreaker extends EventEmitter {
       this.emit('cacheMiss');
     }
 
-    if (!this[ENABLED]) {
+    if (!hasFlag(this, kStateEnabled)) {
       const result = this.action.apply(context, args);
       return (typeof result.then === 'function')
         ? result
@@ -644,7 +671,7 @@ class CircuitBreaker extends EventEmitter {
       return fallback(this, error, args) ||
         Promise.reject(error);
     }
-    this[PENDING_CLOSE] = false;
+    this[STATE_BITSET] &= ~kStatePendingClose;
 
     let timeout;
     let timeoutError = false;
@@ -797,7 +824,7 @@ class CircuitBreaker extends EventEmitter {
    * @returns {void}
    */
   enable () {
-    this[ENABLED] = true;
+    this[STATE_BITSET] |= kStateEnabled;
     this.status.startListeneningForRotateEvent();
   }
 
@@ -807,7 +834,7 @@ class CircuitBreaker extends EventEmitter {
    * @returns {void}
    */
   disable () {
-    this[ENABLED] = false;
+    this[STATE_BITSET] &= ~kStateEnabled;
     this.status.removeRotateBucketControllerListener();
   }
 }