--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -281,21 +281,6 @@ class CircuitBreaker extends EventEmitter {
 
     if (options.maxFailures) console.error(deprecation);
 
-    const increment = property =>
-      (result, runTime) => this[STATUS].increment(property, runTime);
-
-    this.on('success', increment('successes'));
-    this.on('failure', increment('failures'));
-    this.on('fallback', increment('fallbacks'));
-    this.on('timeout', increment('timeouts'));
-    this.on('fire', increment('fires'));
-    this.on('reject', increment('rejects'));
-    this.on('cacheHit', increment('cacheHits'));
-    this.on('cacheMiss', increment('cacheMisses'));
-    this.on('open', _ => this[STATUS].open());
-    this.on('close', _ => this[STATUS].close());
-    this.on('semaphoreLocked', increment('semaphoreRejections'));
-
     /**
      * @param {CircuitBreaker} circuit This current circuit
      * @returns {function(): void} A bound reset callback
@@ -378,6 +363,7 @@ class CircuitBreaker extends EventEmitter {
        * Emitted when the breaker is reset allowing the action to execute again
        * @event CircuitBreaker#close
        */
+      this[STATUS].close();
       this.emit('close');
     }
   }
@@ -400,6 +386,7 @@ class CircuitBreaker extends EventEmitter {
        * failure percentage greater than `options.errorThresholdPercentage`.
        * @event CircuitBreaker#open
        */
+      this[STATUS].open();
       this.emit('open');
     }
   }
@@ -585,7 +572,8 @@ class CircuitBreaker extends EventEmitter {
    * @fires CircuitBreaker#semaphoreLocked
    */
   fire (...args) {
-    return this.call(this.action, ...args);
+    args.unshift(this.action);
+    return this.call.apply(this, args);
   }
 
   /**
@@ -628,6 +616,7 @@ class CircuitBreaker extends EventEmitter {
      * @event CircuitBreaker#fire
      * @type {any} the arguments passed to the fired function
      */
+    this[STATUS].increment('fires');
     this.emit('fire', args);
 
     // If cache is enabled, check if we have a cached value
@@ -640,6 +629,7 @@ class CircuitBreaker extends EventEmitter {
          * and finds a value.
          * @event CircuitBreaker#cacheHit
          */
+        this[STATUS].increment('cacheHits');
         this.emit('cacheHit');
         return cached;
       }
@@ -648,6 +638,7 @@ class CircuitBreaker extends EventEmitter {
        * the cache, but the cache option is enabled.
        * @event CircuitBreaker#cacheMiss
        */
+      this[STATUS].increment('cacheMisses');
       this.emit('cacheMiss');
     }
 
@@ -666,6 +657,7 @@ class CircuitBreaker extends EventEmitter {
        */
       const error = buildError('Breaker is open', 'EOPENBREAKER');
 
+      this[STATUS].increment('rejects');
       this.emit('reject', error);
 
       return fallback(this, error, args) ||
@@ -693,6 +685,7 @@ class CircuitBreaker extends EventEmitter {
                * @event CircuitBreaker#timeout
                * @type {Error}
                */
+              this[STATUS].increment('timeouts', latency);
               this.emit('timeout', error, latency, args);
               handleError(error, this, timeout, args, latency, resolve, reject);
               if (this.options.abortController) {
@@ -715,7 +708,9 @@ class CircuitBreaker extends EventEmitter {
                * @event CircuitBreaker#success
                * @type {any} the return value from the circuit
                */
-              this.emit('success', result, (Date.now() - latencyStartTime));
+              const runTime = (Date.now() - latencyStartTime);
+              this[STATUS].increment('successes', runTime);
+              this.emit('success', result, runTime);
               this.semaphore.release();
               resolve(result);
               if (this.options.cache) {
@@ -751,6 +746,7 @@ class CircuitBreaker extends EventEmitter {
          * @event CircuitBreaker#semaphoreLocked
          * @type {Error}
          */
+        this[STATUS].increment('semaphoreRejections', latency);
         this.emit('semaphoreLocked', err, latency);
         handleError(err, this, timeout, args, latency, resolve, reject);
       }
@@ -844,6 +840,7 @@ function handleError (error, circuit, timeout, args, latency, resolve, reject) {
 
   if (circuit.options.errorFilter(error, ...args)) {
     // The error was filtered, so emit 'success'
+    circuit[STATUS].increment('successes', latency);
     circuit.emit('success', error, latency);
   } else {
     // Error was not filtered, so emit 'failure'
@@ -869,6 +866,7 @@ function fallback (circuit, err, args) {
        * @event CircuitBreaker#fallback
        * @type {any} the return value of the fallback function
        */
+      circuit[STATUS].increment('fallbacks', err);
       circuit.emit('fallback', result, err);
       if (result instanceof Promise) return result;
       return Promise.resolve(result);
@@ -884,6 +882,7 @@ function fail (circuit, err, args, latency) {
    * @event CircuitBreaker#failure
    * @type {Error}
    */
+  circuit[STATUS].increment('failures', latency);
   circuit.emit('failure', err, latency, args);
   if (circuit.warmUp) return;