--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -13,6 +13,8 @@ const kStateShutdown = 1 << 5;
 const kStateWarmingUp = 1 << 6;
 const kStateEnabled = 1 << 7;
 
+const kCircuitFire = Symbol('kCircuitFire');
+
 const STATE_BITSET = Symbol('stateBitset');
 const STATE = Symbol('state');
 
@@ -572,8 +574,7 @@ class CircuitBreaker extends EventEmitter {
    * @fires CircuitBreaker#semaphoreLocked
    */
   fire (...args) {
-    args.unshift(this.action);
-    return this.call.apply(this, args);
+    return this[kCircuitFire](this.action, args);
   }
 
   /**
@@ -601,12 +602,14 @@ class CircuitBreaker extends EventEmitter {
    * @fires CircuitBreaker#semaphoreLocked
    */
   call (context, ...rest) {
+    return this[kCircuitFire](context, rest);
+  }
+
+  [kCircuitFire] (context, args) {
     if (this.isShutdown) {
       const err = buildError('The circuit has been shutdown.', 'ESHUTDOWN');
       return Promise.reject(err);
     }
-
-    const args = rest.slice();
 
     // Need to create variable here to prevent extra calls if cache is disabled
     let cacheKey = '';
@@ -621,7 +624,7 @@ class CircuitBreaker extends EventEmitter {
 
     // If cache is enabled, check if we have a cached value
     if (this.options.cache) {
-      cacheKey = this.options.cacheGetKey.apply(this, rest);
+      cacheKey = this.options.cacheGetKey.apply(this, args.slice());
       const cached = this.options.cacheTransport.get(cacheKey);
       if (cached) {
         /**