--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -30,6 +30,8 @@ const LAST_TIMER_AT = Symbol('last-timer-at');
 const deprecation = `options.maxFailures is deprecated. \
 Please use options.errorThresholdPercentage`;
 
+const noopErrorFilter = _ => false;
+
 function hasFlag (object, flag) {
   return (object[STATE_BITSET] & flag) > 0;
 }
@@ -171,7 +173,7 @@ class CircuitBreaker extends EventEmitter {
     this.options.capacity = Number.isInteger(options.capacity)
       ? options.capacity
       : Number.MAX_SAFE_INTEGER;
-    this.options.errorFilter = options.errorFilter || (_ => false);
+    this.options.errorFilter = options.errorFilter || noopErrorFilter;
     this.options.cacheTTL = options.cacheTTL ?? 0;
     this.options.cacheGetKey = options.cacheGetKey ??
       ((...args) => JSON.stringify(args));
@@ -841,7 +843,7 @@ class CircuitBreaker extends EventEmitter {
 function handleError (error, circuit, timeout, args, latency, resolve, reject) {
   clearTimeout(timeout);
 
-  if (circuit.options.errorFilter(error, ...args)) {
+  if (circuit.options.errorFilter !== noopErrorFilter && circuit.options.errorFilter(error, ...args)) {
     // The error was filtered, so emit 'success'
     circuit[STATUS].increment('successes', latency);
     circuit.emit('success', error, latency);