--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -215,7 +215,9 @@ class CircuitBreaker extends EventEmitter {
       : 0;
     this[STATE_BITSET] = 0;
 
-    if (options.allowWarmUp === true) { this[STATE_BITSET] |= kStateWarmingUp; }
+    if (options.allowWarmUp === true) {
+      this[STATE_BITSET] |= kStateWarmingUp;
+    }
 
     // The user can pass in a Status object to initialize the Status/stats
     if (this.options.status) {
@@ -281,7 +283,9 @@ class CircuitBreaker extends EventEmitter {
 
     if (typeof action !== 'function') {
       this.action = _ => Promise.resolve(action);
-    } else this.action = action;
+    } else {
+      this.action = action;
+    }
 
     if (options.maxFailures) console.error(deprecation);