--- a/lib/circuit.js
+++ b/lib/circuit.js
@@ -867,9 +867,12 @@ function handleError (error, circuit, timeout, args, latency, resolve, reject) {
 function fallback (circuit, err, args) {
   if (circuit[FALLBACK_FUNCTION]) {
     try {
+      // since this is the last method to be called,
+      // is safe to modify the args
+      args.push(err);
       const result =
-      circuit[FALLBACK_FUNCTION]
-        .apply(circuit[FALLBACK_FUNCTION], [...args, err]);
+        circuit[FALLBACK_FUNCTION]
+          .apply(circuit[FALLBACK_FUNCTION], args);
       /**
        * Emitted when the circuit breaker executes a fallback function
        * @event CircuitBreaker#fallback