--- a/lib/semaphore.js
+++ b/lib/semaphore.js
@@ -1,37 +1,35 @@
 'use strict';
 
-module.exports = exports = semaphore;
+const kSemaphoreCounter = Symbol('kSemaphoreCounter');
+const kSemaphoreResolvers = Symbol('kSemaphoreResolvers');
 
-function semaphore (count) {
-  const resolvers = [];
-  let counter = count;
+class Semaphore {
+  constructor (count) {
+    this[kSemaphoreCounter] = count;
+    this[kSemaphoreResolvers] = [];
+  }
 
-  const sem = {
-    take,
-    release,
-    test
-  };
+  get count () {
+    return this[kSemaphoreCounter];
+  }
 
-  Object.defineProperty(sem, 'count', {
-    get: _ => counter,
-    enumerable: true
-  });
+  take (timeout) {
+    if (this[kSemaphoreCounter] > 0) {
+      --this[kSemaphoreCounter];
 
-  return sem;
-
-  function take (timeout) {
-    if (counter > 0) {
-      --counter;
-      return Promise.resolve(release);
+      return Promise.resolve(this.release.bind(this));
     }
+
     return new Promise((resolve, reject) => {
-      resolvers.push(_ => {
-        --counter;
-        resolve(release);
+      this[kSemaphoreResolvers].push(_ => {
+        --this[kSemaphoreCounter];
+        resolve(this.release.bind(this));
       });
+
       if (timeout) {
+        // TODO: Clean this timeout
         setTimeout(_ => {
-          resolvers.shift();
+          this[kSemaphoreResolvers].shift();
           const err = new Error(`Timed out after ${timeout}ms`);
           err.code = 'ETIMEDOUT';
           reject(err);
@@ -40,15 +38,25 @@ function semaphore (count) {
     });
   }
 
-  function release () {
-    counter++;
-    if (resolvers.length > 0) {
-      resolvers.shift()();
+  release () {
+    this[kSemaphoreCounter]++;
+
+    if (this[kSemaphoreResolvers].length > 0) {
+      this[kSemaphoreResolvers].shift()();
     }
   }
 
-  function test () {
-    if (counter < 1) return false;
-    return take() && true;
+  test () {
+    if (this[kSemaphoreCounter] < 1) {
+      return false;
+    }
+
+    return this.take() && true;
   }
 }
+
+function SemaphoreCtor (count) {
+  return new Semaphore(count);
+}
+
+module.exports = exports = SemaphoreCtor;