--- a/lib/semaphore.js
+++ b/lib/semaphore.js
@@ -13,28 +13,18 @@ class Semaphore {
     return this[kSemaphoreCounter];
   }
 
-  take (timeout) {
+  take () {
     if (this[kSemaphoreCounter] > 0) {
       --this[kSemaphoreCounter];
 
       return Promise.resolve(this.release.bind(this));
     }
 
-    return new Promise((resolve, reject) => {
+    return new Promise(resolve => {
       this[kSemaphoreResolvers].push(_ => {
         --this[kSemaphoreCounter];
         resolve(this.release.bind(this));
       });
-
-      if (timeout) {
-        // TODO: Clean this timeout
-        setTimeout(_ => {
-          this[kSemaphoreResolvers].shift();
-          const err = new Error(`Timed out after ${timeout}ms`);
-          err.code = 'ETIMEDOUT';
-          reject(err);
-        }, timeout);
-      }
     });
   }