--- a/lib/semaphore.js
+++ b/lib/semaphore.js
@@ -41,7 +41,9 @@ class Semaphore {
       return false;
     }
 
-    return this.take() && true;
+    --this[kSemaphoreCounter];
+
+    return true;
   }
 }