--- a/lib/status.js
+++ b/lib/status.js
@@ -10,6 +10,8 @@ const ROTATE_EVENT_NAME = Symbol('rotate-event-name');
 
 const EventEmitter = require('events').EventEmitter;
 
+const bucketKeysForStats = ['failures', 'fallbacks', 'successes', 'rejects', 'fires', 'timeouts', 'cacheHits', 'cacheMisses', 'semaphoreRejections'];
+
 /**
  * Tracks execution status for a given {@link CircuitBreaker}.
  * A Status instance is created for every {@link CircuitBreaker}
@@ -113,12 +115,13 @@ class Status extends EventEmitter {
    */
   get stats () {
     const totals = this[WINDOW].reduce((acc, val) => {
-      if (!val) { return acc; }
-      Object.keys(acc).forEach(key => {
-        if (key !== 'latencyTimes' && key !== 'percentiles') {
-          (acc[key] += val[key] || 0);
-        }
-      });
+      if (!val) {
+        return acc;
+      }
+
+      for (const key of bucketKeysForStats) {
+        acc[key] += val[key] || 0;
+      }
 
       if (this.rollingPercentilesEnabled) {
         if (val.latencyTimes) {