--- a/lib/status.js
+++ b/lib/status.js
@@ -125,7 +125,7 @@ class Status extends EventEmitter {
 
       if (this.rollingPercentilesEnabled) {
         if (val.latencyTimes) {
-          acc.latencyTimes = acc.latencyTimes.concat(val.latencyTimes);
+          Array.prototype.push.apply(acc.latencyTimes, val.latencyTimes);
         }
       }
       return acc;