@H4ad 2024-05-26
Hey @lholmquist, can you take a look at this PR?

If you don't think it's worth those changes, let me know to close this one.